Front: Move 'pages' to 'src' folder
