+++ front/src/pages/404.tsx
