+++ front/src/pages/_app.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import type { AppProps } from "next/app";
 import { QueryClient, QueryClientProvider, Hydrate, QueryCache } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
-import "../styles/global.css";
+import "../../styles/global.css";
 import { Box } from "@mui/material";
-import MeeloAppBar from "../src/components/appbar/appbar";
+import MeeloAppBar from "../components/appbar/appbar";
 import { ErrorBoundary } from 'react-error-boundary'
 import toast, { Toaster } from 'react-hot-toast';
